+++ CHANGELOG.md
@@ -18,7 +18,8 @@
 - **command:** `SsvCommand/SsvCommandRef` strict typings
 - **command:** `CommandAsync` has been removed - use `command` instead to create instance or use type `Command`
 - **command:** `ssvCommandParams` and commandCreator params must be in defined in `[]` e.g. `[hero]`
-- **command:** `ssvCommandRef` `command` is now as signal
+- **command:** `ssvCommand` `command` is now signal
+- **command:** `ssvCommandRef` `command` is now signal
 
 ## 4.0.0 (2025-11-14)